Refactors request builder and hasprovider trait
